fix: long dates and lesson date for homework

+++ Data_Tier/src/main/java/sep3/project/data_tier/entity/LessonEntity.java
@@ -15,7 +15,7 @@ public class LessonEntity {
 	@UuidGenerator
 	private String id;
 	@Column
-	private int date;
+	private long date;
 	@Column
 	private String topic;
 	@Column
@@ -41,7 +41,7 @@ public class LessonEntity {
 	public LessonEntity() {
 	}
 
-	public LessonEntity(int date, String topic, String description) {
+	public LessonEntity(long date, String topic, String description) {
 		this.date = date;
 		this.topic = topic;
 		this.description = description;
@@ -55,11 +55,11 @@ public class LessonEntity {
 		this.id = id;
 	}
 
-	public int getDate() {
+	public long getDate() {
 		return date;
 	}
 
-	public void setDate(int date) {
+	public void setDate(long date) {
 		this.date = date;
 	}
